sagemath: change do_install()
The new way installs by whitelist, only what is necessary (it's possible
that something else can be removed in src, but not a lot since src/sage
is required for the internal help)

Also add two more patches: the last remaining issue on i686 and a patch
that avoids the need to install the build directory.

Hopefully now all tests pass on x86_64, x86_64-musl, and i686.  Both
when running in the build chroot by xbps-src check, and also when
running after installation by sage --testall.